+++ file_processing/__init__.py
@@ -1,27 +1,4 @@
 from .file import File
 from .directory import Directory
-from .txt_processor import TextFileProcessor
-from .pdf_processor import PdfFileProcessor
-from .docx_processor import DocxFileProcessor
-from .ocr_decorator import OCRDecorator
-from .msg_processor import MsgFileProcessor
-from .png_processor import PngFileProcessor
-from .xlsx_processor import XlsxFileProcessor
-from .pptx_processor import PptxFileProcessor
-from .rtf_processor import RtfFileProcessor
-from .html_processor import HtmlFileProcessor
-from .xml_processor import XmlFileProcessor
-from .jpeg_processor import JpegFileProcessor
-from .csv_processor import CsvFileProcessor
-from .json_processor import JsonFileProcessor
-from .zip_processor import ZipFileProcessor
-from .audio_processor import AudioFileProcessor
-from .transcription_decorator import TranscriptionDecorator
-from .generic_processor import GenericFileProcessor
-from .py_processor import PyFileProcessor
-from .gif_processor import GifFileProcessor
-from .tiff_processor import TiffFileProcessor
-from .heic_processor import HeicFileProcessor
-from .file_processor_strategy import FileProcessorStrategy
 
 __all__ = ['File', 'Directory']
