+++ file_processing/processors/audio_processor.py
@@ -6,8 +6,8 @@ from mutagen.oggvorbis import OggVorbis
 from mutagen.aiff import AIFF
 from mutagen.wave import WAVE
 from mutagen.mp4 import MP4
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 class AudioFileProcessor(FileProcessorStrategy):
     def __init__(self, file_path: str, open_file: bool = True) -> None:
