+++ file_processing/processors/csv_processor.py
@@ -1,7 +1,7 @@
 import csv
 import chardet
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
 class CsvFileProcessor(FileProcessorStrategy):
