+++ file_processing/processors/gif_processor.py
@@ -1,9 +1,9 @@
 from PIL import Image
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
 
 
-class PngFileProcessor(FileProcessorStrategy):
+class GifFileProcessor(FileProcessorStrategy):
     def __init__(self, file_path: str, open_file: bool = True) -> None:
         super().__init__(file_path, open_file)
         self.metadata = {
@@ -22,6 +22,8 @@ class PngFileProcessor(FileProcessorStrategy):
                 'mode': image.mode,
                 'width': image.width,
                 'height': image.height,
+                'animated': image.is_animated,
+                'frames': image.n_frames
             })
         except Exception as e:
             raise FileProcessingFailedError(
@@ -30,9 +32,8 @@ class PngFileProcessor(FileProcessorStrategy):
     def save(self, output_path: str = None) -> None:
         try:
             image = Image.open(self.file_path)
-
             save_path = output_path or self.file_path
-            image.save(save_path)
+            image.save(save_path, save_all=True)
         except Exception as e:
             raise FileProcessingFailedError(
                 f"Error encountered while saving to {save_path}: {e}")
