+++ file_processing/processors/json_processor.py
@@ -1,8 +1,8 @@
 import json
 from json.decoder import JSONDecodeError
 import chardet
-from file_processing.file_processor_strategy import FileProcessorStrategy
-from file_processing.errors import FileProcessingFailedError, FileCorruptionError
+from file_processing.tools import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError, FileCorruptionError
 
 
 class JsonFileProcessor(FileProcessorStrategy):
