+++ file_processing/processors/msg_processor.py
@@ -1,6 +1,6 @@
 import extract_msg
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
 class MsgFileProcessor(FileProcessorStrategy):
