+++ file_processing/processors/pdf_processor.py
@@ -1,6 +1,6 @@
 from pypdf import PdfReader, PdfWriter
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
 class PdfFileProcessor(FileProcessorStrategy):
