+++ file_processing/processors/py_processor.py
@@ -1,7 +1,7 @@
 import ast
 import shutil
-from file_processing.file_processor_strategy import FileProcessorStrategy
-from file_processing.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
 
 
 class PyFileProcessor(FileProcessorStrategy):
