+++ file_processing/processors/rtf_processor.py
@@ -1,6 +1,6 @@
 from striprtf.striprtf import rtf_to_text
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
 class RtfFileProcessor(FileProcessorStrategy):
