+++ file_processing/processors/tiff_processor.py
@@ -1,12 +1,9 @@
 from PIL import Image
-from pillow_heif import register_heif_opener
-from file_processing.file_processor_strategy import FileProcessorStrategy
-from file_processing.errors import FileProcessingFailedError
-
-register_heif_opener()  # Supports .heic, .heics, .heif, .heifs, .hif
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
-class HeicFileProcessor(FileProcessorStrategy):
+class TiffFileProcessor(FileProcessorStrategy):
     def __init__(self, file_path: str, open_file: bool = True) -> None:
         super().__init__(file_path, open_file)
         self.metadata = {
@@ -24,7 +21,7 @@ class HeicFileProcessor(FileProcessorStrategy):
                 # mode defines type and width of a pixel (https://pillow.readthedocs.io/en/stable/handbook/concepts.html#concept-modes)
                 'mode': image.mode,
                 'width': image.width,
-                'height': image.height,
+                'height': image.height
             })
         except Exception as e:
             raise FileProcessingFailedError(
