+++ file_processing/processors/xlsx_processor.py
@@ -1,8 +1,8 @@
 from io import BytesIO
 from openpyxl import load_workbook
 import msoffcrypto
-from file_processing.errors import FileProcessingFailedError, FileCorruptionError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError, FileCorruptionError
+from file_processing.tools import FileProcessorStrategy
 
 
 class XlsxFileProcessor(FileProcessorStrategy):
