+++ file_processing/processors/xml_processor.py
@@ -1,12 +1,9 @@
 import chardet
-from file_processing.file_processor_strategy import FileProcessorStrategy
-from file_processing.errors import FileProcessingFailedError
-
-# Specialized python library for pulling html/xml data...
-# https://beautiful-soup-4.readthedocs.io/en/latest/
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
-class HtmlFileProcessor(FileProcessorStrategy):
+class XmlFileProcessor(FileProcessorStrategy):
     def __init__(self, file_path: str, open_file: bool = True) -> None:
         super().__init__(file_path, open_file)
         self.metadata = {
@@ -34,7 +31,7 @@ class HtmlFileProcessor(FileProcessorStrategy):
             })
         except Exception as e:
             raise FileProcessingFailedError(
-                f"Error encountered while processing: {e}")
+                f"Error encountered while processing {self.file_path}: {e}")
 
     def save(self, output_path: str = None) -> None:
         try:
@@ -43,4 +40,4 @@ class HtmlFileProcessor(FileProcessorStrategy):
                 f.write(self.metadata['text'])
         except Exception as e:
             raise FileProcessingFailedError(
-                f"Error encountered while saving: {e}")
+                f"Error encountered while saving {self.file_path}: {e}")
