+++ file_processing/processors/zip_processor.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 import shutil
 import zipfile
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file_processor_strategy import FileProcessorStrategy
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing.tools import FileProcessorStrategy
 
 
 class ZipFileProcessor(FileProcessorStrategy):
