+++ file_processing/tools/transcription_decorator.py
@@ -1,6 +1,6 @@
 import whisper
-from file_processing.file_processor_strategy import FileProcessorStrategy
-from file_processing.errors import TranscriptionProcessingError
+from file_processing.tools import FileProcessorStrategy
+from file_processing.tools.errors import TranscriptionProcessingError
 
 
 class TranscriptionDecorator:
