+++ tests/test_audio_processor.py
@@ -7,8 +7,8 @@ from mutagen.mp3 import MP3
 from mutagen.flac import FLAC
 from mutagen.oggvorbis import OggVorbis
 from mutagen.mp4 import MP4
-from file_processing.errors import FileProcessingFailedError
-from file_processing.file import File
+from file_processing.tools.errors import FileProcessingFailedError
+from file_processing import File
 
 
 variable_names = "path, bitrate, length, artist, date, title, organization"
