+++ tests/test_msg_processor.py
@@ -2,8 +2,8 @@ import os
 from unittest.mock import patch
 from datetime import datetime
 import pytest
-from file_processing.file import File
-from file_processing.errors import FileProcessingFailedError
+from file_processing import File
+from file_processing.tools.errors import FileProcessingFailedError
 
 
 variable_names = "path, text_length, subject, date, sender"
