+++ tests/test_ocr_decorator.py
@@ -1,7 +1,7 @@
 import re
 import pytest
-from file_processing.file import File
-from file_processing.errors import NotOCRApplciableError
+from file_processing import File
+from file_processing.tools.errors import NotOCRApplciableError
 
 PDF_SAMPLES = ["tests/resources/test_files/test_ocr_text.pdf", "tests/resources/test_files/test_ocr_text_2.pdf"]
 JPEG_SAMPLES = ["tests/resources/test_files/test_ocr_text.jpg", "tests/resources/test_files/test_ocr_text_2.jpg"]
