+++ tests/test_py_processor.py
@@ -1,7 +1,7 @@
 from unittest.mock import patch
 import pytest
-from file_processing.file import File
-from file_processing.errors import FileProcessingFailedError
+from file_processing import File
+from file_processing.tools.errors import FileProcessingFailedError
 
 
 variable_names = "path, num_lines, num_functions, num_classes, num_imports, num_docstrings"
