+++ tests/test_transcription_decorator.py
@@ -1,7 +1,7 @@
 import pytest
 from utils.cosine_similarity import CosineSimilarity
-from file_processing.file import File
-from file_processing.errors import NotTranscriptionApplicableError
+from file_processing import File
+from file_processing.tools.errors import NotTranscriptionApplicableError
 
 variable_names = "path, actual_transcription_path, language"
 values = [
